# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API pour le jeu Shifumi (Pierre-Papier-Ciseaux).",
        "title": "Shifumi API",
        "contact": {
            "name": "Mihajasoa",
            "email": "mihajasoaalain85@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/game": {
            "post": {
                "description": "Crée une nouvelle partie de Shifumi et renvoie les détails de la partie créée.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Créer une nouvelle partie",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shifumi_internal_models.Game"
                        }
                    }
                }
            }
        },
        "/game/{id}": {
            "get": {
                "description": "Récupère les détails d'une partie de Shifumi en fonction de son ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Obtenir les détails d'une partie",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la partie",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shifumi_internal_models.Game"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/game/{id}/join": {
            "post": {
                "description": "Permet à un joueur de rejoindre une partie existante en fournissant un nom d'utilisateur.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Rejoindre une partie existante",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la partie",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Requête pour rejoindre une partie",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/shifumi_internal_models.JoinGameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shifumi_internal_models.Game"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/game/{id}/play": {
            "post": {
                "description": "Permet à un joueur de jouer un tour de Shifumi en fournissant son choix (pierre, papier ou ciseaux).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Jouer un tour de Shifumi",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la partie",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Requête pour jouer un tour",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/shifumi_internal_models.PlayRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shifumi_internal_models.Game"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Renvoie un statut 200 si le service Shifumi est opérationnel.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Vérifier la santé de l'API",
                "responses": {
                    "200": {
                        "description": "Exemple: {\"message\": \"work well\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "shifumi_internal_models.Choice": {
            "type": "string",
            "enum": [
                "rock",
                "paper",
                "scissors"
            ],
            "x-enum-varnames": [
                "Rock",
                "Paper",
                "Scissors"
            ]
        },
        "shifumi_internal_models.Game": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/shifumi_internal_models.Player"
                    }
                },
                "status": {
                    "$ref": "#/definitions/shifumi_internal_models.GameStatus"
                }
            }
        },
        "shifumi_internal_models.GameStatus": {
            "type": "string",
            "enum": [
                "waiting",
                "ready",
                "playing"
            ],
            "x-enum-varnames": [
                "Waiting",
                "Ready",
                "Playing"
            ]
        },
        "shifumi_internal_models.JoinGameRequest": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                }
            }
        },
        "shifumi_internal_models.PlayRequest": {
            "type": "object",
            "properties": {
                "choice": {
                    "$ref": "#/definitions/shifumi_internal_models.Choice"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "shifumi_internal_models.Player": {
            "type": "object",
            "properties": {
                "choice": {
                    "$ref": "#/definitions/shifumi_internal_models.Choice"
                },
                "score": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}